dream/requirements.txt
django==5.0.2
djangorestframework==3.14.0
groq==0.18.0
python-dotenv==1.0.1